+++ dsr_agents/include/dsr_agents/tf_agent.hpp
@@ -24,7 +24,7 @@
 
 // DSR
 #include "dsr/api/dsr_api.h"
-#include "dsr_agents/agent_node.hpp"
+#include "dsr_util/agent_node.hpp"
 
 class TFAgent: public AgentNode{
 	public: